extract error boundry
