+++ dev/app.tsx
@@ -7,42 +7,9 @@ import React, {
 } from 'react'
 import {createRoot} from 'react-dom/client'
 import {Img, useImage} from '../src/index'
+import {ErrorBoundary} from './ErrorBoundry'
 
 navigator.serviceWorker.register('/sw.js', {scope: './'})
-
-interface ErrorBoundary {
-  props: {
-    children: React.ReactNode
-    onError?: React.ReactNode
-  }
-}
-class ErrorBoundary extends React.Component implements ErrorBoundary {
-  state: {
-    hasError: boolean
-  }
-  onError: React.ReactNode
-
-  constructor(props) {
-    super(props)
-    this.state = {hasError: false}
-    this.onError = props.onError
-  }
-
-  static getDerivedStateFromError(error) {
-    // Update state so the next render will show the fallback UI.
-    return {hasError: error}
-  }
-
-  render() {
-    if (this.state.hasError) {
-      if (this.onError) return this.onError
-      // You can render any custom fallback UI
-      return <h1>Something went wrong.</h1>
-    }
-
-    return this.props.children
-  }
-}
 
 const randSeconds = (min, max) =>
   Math.floor(Math.random() * (max - min + 1) + min)
@@ -334,7 +301,6 @@ function App() {
               unloader={<div>❎ test failed</div>}
             />
           </div>
-
           <div className="testCase">
             <h3>Should not show anything</h3>
             <Img
@@ -343,7 +309,6 @@ function App() {
               unloader={<div>✅ test passed</div>}
             />
           </div>
-
           <div className="testCase">
             <h3>Should show unloader</h3>
             <Img
@@ -353,11 +318,9 @@ function App() {
               unloader={<div>✅ test passed</div>}
             />
           </div>
-
           <div className="testCase">
             <ChangeSrc renderId={renderId} />
           </div>
-
           <div className="testCase">
             <h3>Suspense</h3>
             <Timer delay={rand2} />
@@ -371,7 +334,6 @@ function App() {
               </Suspense>
             </ErrorBoundary>
           </div>
-
           <div className="testCase">
             <h3>Suspense wont load</h3>
             <ErrorBoundary onError={<div>✅ test passed</div>}>
@@ -384,11 +346,9 @@ function App() {
               </Suspense>
             </ErrorBoundary>
           </div>
-
           <div className="testCase">
             <ReuseCache renderId={renderId} />
           </div>
-
           <div className="testCase">
             <div>
               <ErrorBoundary>
@@ -400,7 +360,6 @@ function App() {
               </ErrorBoundary>
             </div>
           </div>
-
           <div className="testCase">
             <ErrorBoundary>
               <HooksLegacyExample rand={rand4} />
@@ -422,5 +381,5 @@ function App() {
 const node = document.createElement('div')
 node.id = 'root'
 document.body.appendChild(node)
-const rootElement = document.getElementById('root')
-createRoot(rootElement!).render(<App />)
+const rootElement = document.getElementById('root') as HTMLElement
+createRoot(rootElement).render(<App />)
